i think i get my edge cuts back?
all the revert commits I have done thus far is because layout editor crashes, and Event Viewer pointed to "NVIDIA OpenGL Driver" in each instance, and then both schematic and layout editors would crash when launching. Went to NVIDIA website to install the driver, selecting the option to perform a clean install in the wizard.

PSA: the express installation option in the driver installer and GeForce Experience leave a lot of duplicate files for programs to find and incorrectly use, make sure to perform a clean install every time you update NVIDIA drivers