preliminary work on USB hub
The Texas Instruments USB hub that is currently in there seems like the ticket, but I am still not sure about how USB C connectors will work with the hub IC since it is only for USB 2.0. The main issue is that USB 3.0 and up natively support 3 different power delivery modes up to 15W (and USB power delivery supporting up to 100W), but this is higher than USB 2.0 capabilities and I don't want to fry the thing.